lambda_event_ai/requirements.txt
av==14.4.0
numpy==1.26.4
pillow==11.3.0
boto3==1.39.2
imageio[pyav]
